% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{add_blob}
\alias{add_blob}
\title{helper to add patch (blob) to existing plot.}
\usage{
add_blob(longdat, center = c(-150, 15), sigma = NULL, fac = 60, rotate = 0)
}
\arguments{
\item{longdat}{Data.}

\item{center}{lon, lat of center of the blob.}

\item{sigma}{Covariance of Gaussian blob.}

\item{fac}{How strong should this blob be.}

\item{rotate}{how much to rotate the blob.}
}
\value{
The resulting long dataset.
}
\description{
helper to add patch (blob) to existing plot.
}
