% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_checkerboard}
\alias{add_checkerboard}
\title{Takes in a (2^k x 2^k) matrix, and adds a checkerboard pattern whose squares
are sized \code{div}, by adding and minusing a value \code{sig} on the
alternating checkboard colors.}
\usage{
add_checkerboard(M, div, sig = 1)
}
\arguments{
\item{M}{original matrix to apply checkerboard pattern to.}

\item{div}{some power of 2 (i.e. 2^0, 2^1, ...), which is the size of the
checkerboard square you want to create.}

\item{sig}{how much to add and subgract, in applying the checkerboard
pattern.}
}
\value{
A matrix of the same size as the input \code{M}, but with a
  checkerboard pattern mean added to it.
}
\description{
Takes in a (2^k x 2^k) matrix, and adds a checkerboard pattern whose squares
are sized \code{div}, by adding and minusing a value \code{sig} on the
alternating checkboard colors.
}
