% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_dimnames}
\alias{add_dimnames}
\title{Helper to add dimension names to matrix \code{M}.}
\usage{
add_dimnames(M)
}
\arguments{
\item{M}{Matrix that contains image.}
}
\description{
Helper to add dimension names to matrix \code{M}.
}
