% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{base_r_plot}
\alias{base_r_plot}
\title{Base R plot for plotting latitude and longitude.}
\usage{
base_r_plot(
  lon,
  lat,
  maxlat,
  val,
  colfun = NULL,
  maxcol = NULL,
  cex_dat = 3,
  ...
)
}
\arguments{
\item{lon}{Longitude.}

\item{lat}{Latitude.}

\item{maxlat}{Maximum latitude value.}

\item{val}{Values to plot.}

\item{colfun}{A function that takes in a numeric vector of numbers between 0
and 1, and produces a vector of colors, to use for plotting the grid.}

\item{maxcol}{Maximum value to use for color evaluation}
}
\description{
Base R plot for plotting latitude and longitude.
}
