% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{binmat}
\alias{binmat}
\title{Nonoverlapping matrix average.}
\usage{
binmat(m, window)
}
\arguments{
\item{m}{matrix}

\item{window}{(window x window) sized non-overlapping bins.}
}
\value{
Matrix of same size but with non-overlapping average.
}
\description{
Nonoverlapping matrix average.
}
\examples{
\dontrun{
m = matrix(rnorm(100), nrow=10, ncol=10)
stopifnot(all(m \%>\% bin_matrix(1) == m))
}
}
