% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landcrossing.R
\name{both_start_and_end_are_in_box}
\alias{both_start_and_end_are_in_box}
\title{Helper function that checks if (start) and (end) are in box define by
(latrange) and (lonrange).}
\usage{
both_start_and_end_are_in_box(start, end, latrange, lonrage)
}
\description{
Helper function that checks if (start) and (end) are in box define by
(latrange) and (lonrange).
}
