% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landcrossing.R
\name{check_angles}
\alias{check_angles}
\title{In a triangle with base a=(x0, y0) and b=(x1, y1), is c=|land_coord| in the
"middle" i.e. are both angles (c-b-a) and (c-a-b) smaller than 90 degrees?}
\usage{
check_angles(a, b, land_coord)
}
\description{
In a triangle with base a=(x0, y0) and b=(x1, y1), is c=|land_coord| in the
"middle" i.e. are both angles (c-b-a) and (c-a-b) smaller than 90 degrees?
}
