% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landcrossing.R
\name{check_cross_land}
\alias{check_cross_land}
\title{Checks if two (lon, lat) points -- (x0, y0) and (x1, y1) -- crosses any land
in \code{land_coords}.}
\usage{
check_cross_land(x0, y0, x1, y1, land_coords, maxdist)
}
\arguments{
\item{x0}{x0}

\item{y0}{y0}

\item{x1}{x1}

\item{y1}{y1}

\item{land_coords}{A data frame whose names are \code{lon} and \code{lat}}

\item{maxdist}{The maximum distance allowed between any land coordinates}
}
\description{
Checks if two (lon, lat) points -- (x0, y0) and (x1, y1) -- crosses any land
in \code{land_coords}.
}
