% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landcrossing.R
\name{check_cross_land_outer2}
\alias{check_cross_land_outer2}
\title{This does the same thing as \code{check_cross_land_outer()} without the
mclapply, which seems to be wacky.}
\usage{
check_cross_land_outer2(
  coords,
  land_coords,
  lonrange = NULL,
  latrange = NULL,
  verbose = FALSE
)
}
\description{
This does the same thing as \code{check_cross_land_outer()} without the
mclapply, which seems to be wacky.
}
