% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omd.R
\name{check_sinkhorn_inputs}
\alias{check_sinkhorn_inputs}
\title{Basic checks for the sinkhorn inputs.}
\usage{
check_sinkhorn_inputs(costm, lambda, invec, outvec)
}
\arguments{
\item{lambda}{Regularization parameter; large value is slow but gives
solution closer to the exact Wasserstein distance.}

\item{invec}{Input vector (size m)}

\item{outvec}{Output vector (size n)}
}
\description{
Basic checks for the sinkhorn inputs.
}
