% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landcrossing.R
\name{complete_rectangle}
\alias{complete_rectangle}
\title{Takes a data matrix with columns: lat, lon, val. Then, it makes the
grid a complete rectangle; more specifically, it adds new rows with lat/lon
values that were missing in \code{dat}, and adds rows with empty (NA) values
to the matrix.}
\usage{
complete_rectangle(dat)
}
\arguments{
\item{dat}{Data matrix.}
}
\value{
Completed data
}
\description{
Takes a data matrix with columns: lat, lon, val. Then, it makes the
grid a complete rectangle; more specifically, it adds new rows with lat/lon
values that were missing in \code{dat}, and adds rows with empty (NA) values
to the matrix.
}
