% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landcrossing.R
\name{complete_rectangle_barebones}
\alias{complete_rectangle_barebones}
\title{Same as \code{complete_rectangle} but doesn't need \code{month} or
\code{dat_type},.}
\usage{
complete_rectangle_barebones(dat)
}
\description{
Same as \code{complete_rectangle} but doesn't need \code{month} or
\code{dat_type},.
}
