% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compute_wasserstein}
\alias{compute_wasserstein}
\title{Compute the wasserstein distance using the algorithm by Aurenhammer,
Hoffmann and Aronov (1998) for finding optimal transference plans in terms
of the squared Euclidean distance in two dimensions.}
\usage{
compute_wasserstein(p1, p2)
}
\arguments{
\item{p1}{The first \code{pgrid} object.}

\item{p2}{The second \code{pgrid} object.}
}
\value{
The Wasserstein distance
}
\description{
Compute the wasserstein distance using the algorithm by Aurenhammer,
Hoffmann and Aronov (1998) for finding optimal transference plans in terms
of the squared Euclidean distance in two dimensions.
}
