% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landcrossing.R
\name{crosses_land}
\alias{crosses_land}
\title{Checks if the line between (start ---- end) crosses any land.}
\usage{
crosses_land(start, end, latrange, lonrange, land_coords)
}
\description{
Checks if the line between (start ---- end) crosses any land.
}
