% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{drawmat_3d}
\alias{drawmat_3d}
\title{Drawing a 3d plot of a 2d matrix, where the Z values are the pixel values.}
\usage{
drawmat_3d(M, zfac = 30)
}
\arguments{
\item{M}{Matrix.}

\item{zfac}{Factor for z limits for the 3d plots.}
}
\description{
Drawing a 3d plot of a 2d matrix, where the Z values are the pixel values.
}
