% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fill_na}
\alias{fill_na}
\title{Fill in NA from surrounding values.}
\usage{
fill_na(mat)
}
\arguments{
\item{mat}{A matrix with possibly missing values}
}
\value{
The same object.
}
\description{
Fill in NA from surrounding values.
}
