% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landcrossing.R
\name{get_all_in_btw_points}
\alias{get_all_in_btw_points}
\title{Get all in between points (that touch the grid).}
\usage{
get_all_in_btw_points(x0, y0, x1, y1, one_unit)
}
\arguments{
\item{x0}{Starting lon.}

\item{y0}{Starting lat.}

\item{x1}{Ending lon.}

\item{y1}{Ending lat.}

\item{one_unit}{Length of one unit, in lon/lat.}
}
\value{
All points in between that touch the grid (two-column matrix).
}
\description{
Get all in between points (that touch the grid).
}
