% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_matrix.R
\name{get_costm}
\alias{get_costm}
\title{Given image |mat| whose number of entries is N, calculate cost matrix (of
size (M x M).}
\usage{
get_costm(mat, land_inds)
}
\arguments{
\item{mat}{Input matrix (image).}

\item{land_inds}{Indices of |mat| that correspond to land.}
}
\value{
Cost matrix of size (M x M).
}
\description{
Given image |mat| whose number of entries is N, calculate cost matrix (of
size (M x M).
}
