% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landcrossing.R
\name{get_land_coords}
\alias{get_land_coords}
\title{From a long data frame with lat, lon, obtain the land indices by}
\usage{
get_land_coords(dat)
}
\arguments{
\item{longdat}{Long data matrix containing \code{lat}, \code{lon}.}
}
\value{
Subset of the rows of \code{dat} that are land.
}
\description{
(1) completing the rectangle, and
}
\details{
(2) marking the overlaps with land using \code{mark_land()}.
}
