% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_time_space_box}
\alias{get_time_space_box}
\title{Takes data that contains "month", "lon" and "lat" as columns, and isolates
it to one month \code{mo} and to a latitude/longitutde range \code{lonrange}
and \code{latrange}.}
\usage{
get_time_space_box(dat, mo, latrange, lonrange, fill_na = FALSE)
}
\arguments{
\item{dat}{data matrix. Must include "lat" "lon" and "month"}

\item{mo}{Month.}

\item{latrange}{Latitude range.}

\item{lonrange}{Longitude range.}
}
\description{
Takes data that contains "month", "lon" and "lat" as columns, and isolates
it to one month \code{mo} and to a latitude/longitutde range \code{lonrange}
and \code{latrange}.
}
