% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omd.R
\name{image_to_long_format}
\alias{image_to_long_format}
\title{Helper function to transform matrix of pixel values (image) to a matrix of
long format, where one row is one data point.}
\usage{
image_to_long_format(img)
}
\arguments{
\item{img}{matrix of pixel values; the row names and col names are assumed
to contain lat and lon points.}
}
\value{
Matrix with three columns, named `lat`, `lon`, and `val`.
}
\description{
Helper function to transform matrix of pixel values (image) to a matrix of
long format, where one row is one data point.
}
