% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landcrossing.R
\name{intersect_latlon}
\alias{intersect_latlon}
\title{Make the two data sources share the same lat/lon points; discard the
non-overlapping ones. Basically erases a-b from a.}
\usage{
intersect_latlon(a, b)
}
\description{
Make the two data sources share the same lat/lon points; discard the
non-overlapping ones. Basically erases a-b from a.
}
