% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omd.R
\name{long_format_to_image}
\alias{long_format_to_image}
\title{The reverse of \code{image_to_long_format()}.}
\usage{
long_format_to_image(dat)
}
\arguments{
\item{dat}{A data frame (or matrix) whose columns are named: \code{lon},
\code{lat}, \code{val}.}
}
\value{
An image, with values in pixels where values exist, and NA's
  otherwise.
}
\description{
The reverse of \code{image_to_long_format()}.
}
