% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ma.matrix}
\alias{ma.matrix}
\title{Helper function taken directly from the OLIN package. Takes a sliding window
average i.e. (j,k)'th pixel's value is taken from a square window extending
+-delta out in either direction. When the window of size (delta*2+1) x
(delta*2+1) centered at (j,k) tries to extend to outside of the matrix, it's
not allowed to. This is equal to a local regression of zero'th order.}
\usage{
ma.matrix(X, av = "median", delta = 2, edgeNA = FALSE)
}
\arguments{
\item{X}{Data matrix}

\item{av}{What function to use for summarizing. Defaults to median.}

\item{delta}{Window size.}

\item{edgeNA}{if TRUE, then any pixel whose (delta*2+1) x (delta*2+1) window
extends outside of the matrix is assigned zero.}
}
\description{
Helper function taken directly from the OLIN package. Takes a sliding window
average i.e. (j,k)'th pixel's value is taken from a square window extending
+-delta out in either direction. When the window of size (delta*2+1) x
(delta*2+1) centered at (j,k) tries to extend to outside of the matrix, it's
not allowed to. This is equal to a local regression of zero'th order.
}
