% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_colors}
\alias{make_colors}
\title{Make a vector of colors that are all blue, except for the top \code{prob}
quantile of mass transfers.}
\usage{
make_colors(res, prob = 0.1)
}
\arguments{
\item{res}{An object from \code{transform::transform()}.}

\item{prob}{Defaults to 0.1.}
}
\value{
A vector of colors, mostly blue, except for the top \code{prob}
  quantile.
}
\description{
Make a vector of colors that are all blue, except for the top \code{prob}
quantile of mass transfers.
}
