% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_pgrid}
\alias{make_pgrid}
\title{Makes a pgrid object, from a matrix whose row names and column names contain
*equally spaced* latitudes and longitudes.}
\usage{
make_pgrid(dat)
}
\description{
Makes a pgrid object, from a matrix whose row names and column names contain
*equally spaced* latitudes and longitudes.
}
