% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landcrossing.R
\name{mark_land}
\alias{mark_land}
\title{Given lon and lat values, mark which guys are}
\usage{
mark_land(lon, lat, overreact = FALSE, buffer_in_km = 500)
}
\description{
Given lon and lat values, mark which guys are
}
