% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landcrossing.R
\name{my_basic_setdiff}
\alias{my_basic_setdiff}
\title{More basic version of \code{my_setdiff()}.}
\usage{
my_basic_setdiff(a, b, check_if_a_in_b = TRUE)
}
\description{
More basic version of \code{my_setdiff()}.
}
