% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landcrossing.R
\name{my_setdiff}
\alias{my_setdiff}
\title{Obtain the indices of the set difference between two data frames of lat &
lons. If rows 1,3,6 of \code{complete_grid} are missing in
\code{incomplete_grid}, we find those indices.}
\usage{
my_setdiff(incomplete_grid, complete_grid)
}
\arguments{
\item{complete}{grid A complete data frame of lat & lon columns.}

\item{incomplete}{grid An incomplete data frame with lat & lon columns. All
rows must be contained in \code{complete_grid}.}
}
\value{
Row indices of \code{complete_grid} of the rows that are missing in
  \code{incomplete_grid}.
}
\description{
Obtain the indices of the set difference between two data frames of lat &
lons. If rows 1,3,6 of \code{complete_grid} are missing in
\code{incomplete_grid}, we find those indices.
}
