\name{omd-package}
\Rdversion{1.1}
\alias{omd-package}
\alias{omd}
\docType{package}
\title{
 filter and optimize the molecular descriptors for QSAR
}
\description{
 This package including two useful function, which can be used for filter the molecular descriptors matrix for QSAR. 
}
\details{
\tabular{ll}{
Package: \tab omd\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-09-16\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
omd including two functions, fcons and fhcor, which can delete constant and high correlated molecular descriptors, respectively. 
}
\author{
Bin Ma
Maintainer: Bin Ma<binma01@gmail.com>
}
\references{
Ma B, Chen HH, He Y, Xu JM 2009 Quantitative structure-activity relationship (QSAR) models based on molecular structure and effect sizes ofpolycyclic aromatic hydrocarbons (PAHs) dissipation in rhizosphere. Environmental Pollution (Submitted).
}

