% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omd.R
\name{omd}
\alias{omd}
\title{Calculate Wasserstein's distance between two images (either two matrices
representing 2d pixelized images, or two long data frames with columns
"lon", "lat", and "val"). The data does not need to be normalized. "omd"
stands for Ocean Mover's Distance.}
\usage{
omd(
  M1 = NULL,
  M2 = NULL,
  costm = NULL,
  M1_long = NULL,
  M2_long = NULL,
  coordnames = c("lon", "lat"),
  type = c("transport", "sinkhorn"),
  p = 1,
  sinkhorn_lambda = 30,
  sinkhorn_eps = 1e-15,
  sinkhorn_rel_eps = NULL,
  geodesic = FALSE
)
}
\arguments{
\item{M1}{One matrix containing pixel values (m x n). The row names and
column names of this matrix are assumed to contain longitude and latitude
information. Otherwise, lon/lat values of \code{1:m} and \code{1:n} are
sassumed.}

\item{M2}{Another matrix; see \code{M1}.}

\item{costm}{Cost matrix \code{mm x mn}. If not provided, one is calculated
as Euclidean distance.}

\item{M1_long}{Long matrix, with a column called "val". If provided, this
supercedes \code{M1}.}

\item{M2_long}{Another long matrix, with a column called "val". If provided,
this supercedes \code{M2}.}

\item{coordnames}{Defaults to \code{c("lon", "lat")}.}

\item{type}{Whether to use Sinkhorn approximation, or the \code{R}
\code{transport} package.}

\item{p}{Power of Wasserstein's distance.}

\item{sinkhorn_lambda}{Regularization parameter for Sinkhorn.}

\item{sinkhorn_eps}{Tolerance for Sinkhorn objective value.}

\item{sinkhorn_rel_eps}{Relative tolerance for Sinkhorn objective value.}

\item{geodesic}{Whether to use geodesic distance instead of Euclidean
distance.}
}
\value{
A list object with the results and data, of S3 class "omd".
}
\description{
Calculate Wasserstein's distance between two images (either two matrices
representing 2d pixelized images, or two long data frames with columns
"lon", "lat", and "val"). The data does not need to be normalized. "omd"
stands for Ocean Mover's Distance.
}
