% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_dat}
\alias{plot_dat}
\title{(Simpler, easier projection) Make a multi-panel plot of the data.}
\usage{
plot_dat(
  longdat = NULL,
  mat = NULL,
  valname = NULL,
  standardize = TRUE,
  hide_legend = TRUE,
  add_map = FALSE,
  colours = NULL,
  lonrange = NULL,
  latrange = NULL,
  map_projection = TRUE,
  map_orientation = c(-36.92, 200, 0),
  limits = NULL,
  breaks = waiver(),
  legend_name = waiver()
)
}
\arguments{
\item{longdat}{A matrix with columns: \code{lat}, \code{lon},
\code{val}. Optionally, \code{longdat} can have a column named
\code{dat_type}. Also otionally, provide \code{mat}, which will be
converted to long format, then plotted.}

\item{mat}{Defaults to NULL. Will be converted to long format, then plotted
in a single panel}

\item{valname}{Optionally provide name of the measurement
(e.g. \code{valname = "Chlorophyll"}.}
}
\value{
ggplot object.
}
\description{
(Simpler, easier projection) Make a multi-panel plot of the data.
}
