% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_omd}
\alias{plot_omd}
\title{Base R plotting of the optimal transport. Produces three figures: (1) The
"before" image, (2) the "after" image, and (3) the optimal transports.}
\usage{
plot_omd(
  obj,
  lwd_max = 10,
  colfun = NULL,
  cex = 3,
  maxcol = NULL,
  main1 = NULL,
  main2 = NULL,
  ...
)
}
\arguments{
\item{obj}{An omd object.}

\item{colfun}{A function that takes in a numeric vector of numbers between 0
and 1, and produces a vector of colors, to use for plotting the grid.}

\item{cex}{Size of the squares to plot in a grid.}

\item{maxcol}{Maximum value to use for color evaluation}

\item{main1}{Title of first image.}

\item{main2}{Title of second image.}
}
\description{
Base R plotting of the optimal transport. Produces three figures: (1) The
"before" image, (2) the "after" image, and (3) the optimal transports.
}
