% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{plot_vf}
\alias{plot_vf}
\title{Plots a vector field that characterizes the optimal transport.}
\usage{
plot_vf(jan.dat, feb.dat, res)
}
\arguments{
\item{jan.dat}{one dataset.}

\item{feb.dat}{another dataset}
}
\value{
The transport result.
}
\description{
Plots a vector field that characterizes the optimal transport.
}
