% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data.R
\name{process_data}
\alias{process_data}
\title{Month and lon/lat range; also standardizes by dividing by the entire data
sum (so that all data sums to one).}
\usage{
process_data(dat, month, lonrange, latrange)
}
\arguments{
\item{dat}{Data, obtained from \code{read_data()}.}

\item{month}{Month.}

\item{lonrange}{Longitude range.}

\item{latrange}{Latitude range.}
}
\value{
Processed data
}
\description{
Month and lon/lat range; also standardizes by dividing by the entire data
sum (so that all data sums to one).
}
