% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{range_custom}
\alias{range_custom}
\title{Scales \code{x} into between \code{min} and \code{max}.}
\usage{
range_custom(x, min, max)
}
\description{
Scales \code{x} into between \code{min} and \code{max}.
}
