% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data.R
\name{read_data}
\alias{read_data}
\title{Gets the data (in \url{https://github.com/brorfred/ocean_clustering}).
Resolution of (1/2, 1, 2, 4) degrees.}
\usage{
read_data(
  resolution = c("4", "2", "1", "0.5"),
  type = c("Darwin", "Real"),
  datadir = "/home/shyun/Dropbox/research/usc/ocean-provinces/omd/data"
)
}
\arguments{
\item{datadir}{Directory for data.}
}
\value{
Data that contains month, lat, lon, and some other data columns.
}
\description{
Gets the data (in \url{https://github.com/brorfred/ocean_clustering}).
Resolution of (1/2, 1, 2, 4) degrees.
}
