% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omd.R
\name{sinkhorn}
\alias{sinkhorn}
\title{Calculate Sinkhorn distance.}
\usage{
sinkhorn(costm, lambda, invec, outvec, eps = 0.001, verbose = FALSE)
}
\arguments{
\item{lambda}{Regularization parameter; large value is slow but gives
solution closer to the exact Wasserstein distance.}

\item{invec}{Input vector (size m)}

\item{outvec}{Output vector (size n)}

\item{eps}{Tolerance for stopping the algorithm, in terms of row sums of P.}

\item{verbose}{Set to \code{TRUE} if you want loud code. Defaults to \code{TRUE}}

\item{costmat}{(m x n) Cost matrix}
}
\value{
List containing three things; `dist`, the sinkhorn distance, `costm`
  the original cost matrix, `transports` is an (m x n) matrix of the optimal
  (regularized) transports.
}
\description{
Calculate Sinkhorn distance.
}
