% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{smoothdat}
\alias{smoothdat}
\title{A sliding window, of window size (\code{size} x \code{size}), is applied on
the dataset.}
\usage{
smoothdat(dat, size)
}
\arguments{
\item{dat}{Dataset with lat, lon, val.}
}
\description{
A sliding window, of window size (\code{size} x \code{size}), is applied on
the dataset.
}
