% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{smoothmat}
\alias{smoothmat}
\title{A sliding window average, of window size (\code{size} x \code{size}).}
\usage{
smoothmat(M, size)
}
\arguments{
\item{M}{Matrix.}

\item{size}{Size of sliding window.}
}
\value{
Smoothed matrix of the same size
}
\description{
A sliding window average, of window size (\code{size} x \code{size}).
}
