% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{add_blob}
\alias{add_blob}
\title{Add blob.}
\usage{
add_blob(longdat, center = c(-150, 15), sigma = NULL, fac = 60, rotate = 0)
}
\arguments{
\item{longdat}{Data.}

\item{center}{lon, lat of center of the blob.}
}
\value{
abc
}
\description{
Add blob.
}
