% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{drawmat_precise}
\alias{drawmat_precise}
\title{Taken from Justin's convenience functions..}
\usage{
drawmat_precise(mat, contour = FALSE, ...)
}
\description{
Taken from Justin's convenience functions..
}
