% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omd.R
\name{form_cost_matrix}
\alias{form_cost_matrix}
\title{Converts a matrix with `n` rows and with columns `lat`, `lon`, into a cost
matrix of size (n x n).}
\usage{
form_cost_matrix(dat, pow = 2)
}
\arguments{
\item{Matrix}{whose rows are data points, and columns are named "lat" and
"lon".}
}
\value{
p x p cost matrix.
}
\description{
Converts a matrix with `n` rows and with columns `lat`, `lon`, into a cost
matrix of size (n x n).
}
