% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omd.R
\name{omd}
\alias{omd}
\title{Calculate earthmovers distance between two images. They don't have to add up
to 1.}
\usage{
omd(
  M1,
  M2,
  costm = NULL,
  M1_long = NULL,
  M2_long = NULL,
  type = c("transport", "sinkhorn"),
  sinkhorn_lambda = 30,
  sinkhorn_eps = 1e-15,
  sinkhorn_rel_eps = NULL
)
}
\arguments{
\item{M1}{One matrix containing pixel values (m x m).}

\item{M2}{Another matrix.}

\item{costm}{Cost matrix \code{m^2 x m^2}.}

\item{type}{Defaults to "transport".}
}
\value{
Not written yet.
}
\description{
Calculate earthmovers distance between two images. They don't have to add up
to 1.
}
