% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_omd_ggplot}
\alias{plot_omd_ggplot}
\title{ggplot-based plotting of the optimal transport pattern. \code{plot_dat()} is
used for plotting data.}
\usage{
plot_omd_ggplot(
  obj,
  plot_type = c("one", "four"),
  name_from = NULL,
  name_to = NULL,
  add_map = FALSE,
  classify_quantile = TRUE,
  sample_arrows = FALSE,
  map_projection = TRUE,
  map_orientation = c(-36.92, 200, 0),
  return_arrows = FALSE,
  arrow_range = c(0.001, 2)
)
}
\arguments{
\item{obj}{An "omd" class object.}

\item{plot_type}{One or four plots.}

\item{add_map}{If \code{TRUE}, add a map of the coastline.}

\item{classify_quantile}{If \code{TRUE}, classify into blue and red arrows;
otherwise just use blue arrows whose thickness.}

\item{sample_arrows}{If \code{TRUE}, only plot half of the arrows.}

\item{map_projection}{If \code{TRUE}, project onto the "globe", so the plot
appears curved.}

\item{map_orientation}{Orientation to input to \code{coord_map()} from
\code{dplyr}. \url{https://ggplot2.tidyverse.org/reference/coord_map.html}}

\item{return_arrows}{If \code{TRUE}, return a table containing information
about arrows.}

\item{arrow_range}{Range of arrow thicknesses.}
}
\description{
ggplot-based plotting of the optimal transport pattern. \code{plot_dat()} is
used for plotting data.
}
