% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinkhorn.R
\name{sinkhorn}
\alias{sinkhorn}
\title{Calculate Sinkhorn distance.}
\usage{
sinkhorn(
  costm,
  lambda,
  invec,
  outvec,
  eps = 0.001,
  rel_eps = 1e-08,
  niter = 1e+05
)
}
\arguments{
\item{lambda}{Regularization parameter; large value is slow but gives
solution closer to the exact Wasserstein distance.}

\item{invec}{Input vector (size m)}

\item{outvec}{Output vector (size n)}

\item{eps}{Tolerance for stopping the algorithm, in terms of row sums of P.}

\item{rel_eps}{Relative tolerance.}

\item{costmat}{(m x n) Cost matrix}

\item{verbose}{Set to \code{TRUE} if you want loud code. Defaults to \code{TRUE}}
}
\value{
List containing three things; \code{dist}, the sinkhorn distance,
  \code{costm} the original cost matrix, \code{transports} is an (m x n)
  matrix of the optimal (regularized) transports.
}
\description{
Calculate Sinkhorn distance.
}
